requirements.txt
asgiref==3.5.2
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
# django-crispy-forms==2.1
django-widget-tweaks==1.4.8
# dnspython==2.4.2
email-validator==2.1.0
httpagentparser==1.9.5
httplib2==0.21.0
# Jinja2==3.1.2
# oauthlib==3.2.2
# olefile==0.46
# packaging==23.1
# Paste==3.5.3
# pexpect==4.8.0
# pid==2.2.3
# Pillow==10.1.0
pipenv==2023.2.18
# platformdirs==3.9.1
psycopg==3.1.10
psycopg2-binary==2.9.9
pytz==2020.1
python-decouple==3.8
python-dotenv==0.21.1
six==1.16.0
# SQLAlchemy==1.4.50
sqlparse==0.3.1
# uritemplate==4.1.1
# urllib3==1.26.18
# user-agents==2.2.0
validate-email-address==1
# virtualenv==20.21.1
# virtualenv-clone==0.5.7
# websocket-client==1.3.3
# WTForms==3.0.1
# timeout-decorator==0.5.0
python-dotenv==0.21.1